# doc/openapi.yaml
openapi: 3.0.0
info:
  title: Playzone API
  version: '0.1'
servers:
  - url: 'http://localhost:8000'
paths:
...
  /api/users:
    get:
      summary: Récupérer des utilisateurs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  pages:
                    type: integer
                    minimum: 0
                  page:
                    type: integer
                    minimum: 0
                  size:
                    type: integer
                    minimum: 1
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
      operationId: get-api-users
      security:
        - ADMIN: []
      parameters:
        - schema:
            type: integer
          in: query
          name: page
          description: index de la pagination
          required: true
        - schema:
            type: integer
          in: query
          name: size
          description: taille de la pagination
          required: true
      tags:
        - user
    post:
      summary: Créer un utilisateur
      operationId: post-api-users
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
              description: '/api/users/{userId}'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Created'
              examples: {}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorConflict'
      security:
        - ADMIN: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserFormCreate'
        description: ''
      tags:
        - user
  '/api/users/{userId}':
    parameters:
      - schema:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name: userId
        in: path
        required: true
        description: id d'un utilisateur
    get:
      summary: Récupérer un utilisateur
      tags:
        - user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
          content: {}
      operationId: get-api-users-userId
      security:
        - ADMIN: []
    patch:
      summary: Modifier un utilisateur
      operationId: patch-api-users-userId
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
          content: {}
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorConflict'
      security:
        - ADMIN: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserFormPatch'
      tags:
        - user
    delete:
      summary: Supprimer un utilisateur
      operationId: delete-api-users-userId
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
          content: {}
      security:
        - ADMIN: []
      tags:
        - user
  /api/users-emails:
    get:
      summary: Récupérer des adresses email
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: email
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
      operationId: get-api-users-emails
      security:
        - ADMIN: []
        - USER: []
      parameters:
        - schema:
            type: string
          in: query
          name: email
      description: Récupère des adresses email (maximum 10) commençant par la valeur du paramètre 'email'
      tags:
        - user
    parameters: []
...
  /api/tables:
    parameters: []
    get:
      summary: Récupérer des tables
      tags:
        - table
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  pages:
                    type: integer
                    minimum: 0
                  page:
                    type: integer
                    minimum: 0
                  size:
                    type: integer
                    minimum: 1
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DataTable'
              examples:
                Example 1:
                  value:
                    pages: 0
                    page: 0
                    size: 1
                    data:
                      - id: 01BX5ZZKBKACTAV9WEVGEMMVRY
                        table: DimNace
                        headers:
                          - id: 01BX5ZZKBKACTAV9WEVGEMMVRY
                            name: SKNace
                          - id: 01BX5ZZKBKACTAV9WEVGEMMVRY
                            name: CodeNace
                        log:
                          createdAt: '2019-08-24T14:15:22.000Z'
                          updatedAt: '2019-08-24T14:15:22.000Z'
                          createdBy: user@example.com
                          updatedBy: user@example.com
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
      operationId: get-api-tables
      security:
        - ADMIN: []
        - USER: []
      parameters:
        - schema:
            type: integer
            minimum: 0
          in: query
          name: page
          description: index de la pagination
          required: true
        - schema:
            type: integer
            minimum: 1
          in: query
          name: size
          description: taille de la pagination
          required: true
        - schema:
            type: string
            enum:
              - OWNER
              - SHARED
          in: query
          description: '(facultatif) pour n''afficher que les tables créées par l''utilisateur connecté, ou les tables partagées avec l''utilisateur connecté, sinon affiche toutes les tables si ce champ est omis'
          name: from
    post:
      summary: Créer une table
      operationId: post-api-tables-tableId
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
              description: '/api/tables/{tableId}'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Created'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
      security:
        - ADMIN: []
        - USER: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataTableFormCreate'
            examples:
              Example 1:
                value:
                  table: DimNace
                  headers:
                    - name: SKNace
                    - name: CodeNace
      tags:
        - table
...
  '/api/tables/{tableId}/permissions':
    parameters:
      - schema:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name: tableId
        in: path
        required: true
        description: id d'une table
    get:
      summary: Récupérer les permissions d'accès à une table
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TablePermissions'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
          content: {}
      operationId: get-api-tables-tableId-permissions
      security:
        - ADMIN: []
        - USER: []
      tags:
        - permission
  '/api/tables/{tableId}/permissions/{userEmailBase64}':
    parameters:
      - schema:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name: tableId
        in: path
        required: true
        description: id d'une table
      - schema:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name: userEmailBase64
        in: path
        required: true
        description: email en base64 de l'utilisateur avec qui partager la table
    get:
      summary: Récupérer une permission d'accès à une table
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TablePermission'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
          content: {}
      operationId: get-api-tables-id-permissions
      security:
        - ADMIN: []
        - USER: []
      tags:
        - permission
    post:
      summary: Créer ou modifier une permission d'accès à une table
      operationId: put-api-tables-id-permissions
      responses:
        '204':
          description: No Content
          headers: {}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
      security:
        - ADMIN: []
        - USER: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TablePermission'
      tags:
        - permission
    delete:
      summary: Supprimer une permission d'accès à une table
      operationId: delete-api-tables-id-permissions
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
          content: {}
      security:
        - ADMIN: []
        - USER: []
      tags:
        - permission
...
components:
  schemas:
...
    User:
      title: User
      type: object
      x-examples: {}
      properties:
        id:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
        role:
          type: string
          example: USER
          description: '"USER" ou "ADMIN"'
        isActivated:
          type: boolean
        log:
          $ref: '#/components/schemas/HistoryLog'
      required:
        - id
        - firstName
        - lastName
        - email
        - role
        - isActivated
        - log
    UserFormCreate:
      title: UserFormCreate
      type: object
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        role:
          type: string
          example: USER
          description: '"USER" ou "ADMIN"'
        isActivated:
          type: boolean
      required:
        - firstName
        - lastName
        - email
        - password
        - role
    UserFormPatch:
      title: UserFormPatch
      type: object
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        role:
          type: string
          example: USER
          description: '"USER" ou "ADMIN"'
        isActivated:
          type: boolean
...
    TablePermission:
      title: TablePermission
      type: object
    TablePermissions:
      title: TablePermissions
      type: array
      items:
        type: object
        properties:
          email:
            type: string
            format: email
        required:
          - email
    DataTable:
      title: DataTable
      type: object
      properties:
        id:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        table:
          type: string
        headers:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: 01BX5ZZKBKACTAV9WEVGEMMVRY
              name:
                type: string
            required:
              - id
              - name
        log:
          $ref: '#/components/schemas/HistoryLog'
      required:
        - id
        - table
        - headers
        - log
    DataTableFormCreate:
      title: DataTableFormCreate
      type: object
      properties:
        table:
          type: string
        headers:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
            required:
              - name
      required:
        - table
        - headers
...
    ErrorForm:
      title: ErrorForm
      type: object
      properties:
        status:
          type: integer
          example: 400
        path:
          type: string
          example: /api/endpoint
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: name
              message:
                type: string
                example: required
            required:
              - field
              - message
      required:
        - status
        - path
        - errors
    ErrorConflict:
      title: ErrorConflict
      type: object
      properties:
        status:
          type: integer
          example: 409
        path:
          type: string
          example: /api/endpoint
        error:
          type: string
          example: Data already exists
      required:
        - status
        - path
        - error
    ErrorUnauthorized:
      title: ErrorUnauthorized
      type: object
      properties:
        status:
          type: integer
          example: 401
        path:
          type: string
          example: /api/endpoint
        error:
          type: string
          example: Unauthorized
      required:
        - status
        - path
        - error
    ErrorForbidden:
      title: ErrorForbidden
      type: object
      properties:
        status:
          type: integer
          example: 403
        path:
          type: string
          example: /api/endpoint
        error:
          type: string
          example: Forbidden
      required:
        - status
        - path
        - error
    Created:
      title: Created
      type: object
      properties:
        id:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
      required:
        - id
    HistoryLog:
      title: HistoryLog
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: email
        updatedBy:
          type: string
          format: email
      required:
        - createdAt
        - updatedAt
        - createdBy
        - updatedBy
  securitySchemes:
    ADMIN:
      type: http
      scheme: bearer
      description: ''
    USER:
      type: http
      scheme: bearer
...